Set Level 2 broadcast flag if UDP and the broadcast address is specified. (Daniel Lacroix)
When using destination address 255.255.255.255 (level 2 broadcast), this
patch will just allow broadcast.

This allow for example to send OSC message to all the local Ethernet
network by simply using destination address 255.255.255.255.

This patch does not allow level 3 (IP) broadcast (ex: 192.168.0.255
for the network 192.168.0.0/24), because IP broadcast is not a simple
problem and when activated, if some are not correctly using broadcast
and netmask it can do broadcast when point to point was intended.


git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@102 8a0476ba-7714-0410-bd95-fb3ea4a72f84