Add support for multicast groups to LibLo. (SS)
For sending, it adds a "ttl" variable to the lo_address structure, and
uses setsockopt() to set it before sendto() is called.

For receiving, it adds new method for specifying the multicast group
when creating a server.  This group is then joined before the socket
is bound.

New internal functions:
- lo_server_new_with_proto_internal()

New hi-level functions:
- lo_server_thread_new_multicast()

New low-level functions:
- lo_server_new_multicast()
- lo_address_set_ttl()
- lo_address_get_ttl()


git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@109 8a0476ba-7714-0410-bd95-fb3ea4a72f84