Expose the lo_server_dispatch_data() function as public.
Now that we have the serialise()/deserialise() functions, liblo can be
used with raw blocks of memory.  However, once memory has been
deserialised into a lo_message, it can be sent to a destination, but
not handled by a local server.  This patch exposes the steps
surrounding the use of deserialise(), where a block of memory is both
deserialised() _and_ dispatched or queued by the server.  It could be
useful, for example, to make use of liblo as an OSC handling mechanism
for transport layers not supported by liblo natively.

git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@129 8a0476ba-7714-0410-bd95-fb3ea4a72f84