Make the TCP protocol retain session information for multiple connections.
The lo_server.socket member becomes an array of socket identifiers,
where the zeroeth socket is the listening socket.  When poll() is
used, they are pollfd structures, otherwise an anonymous struct with a
.fd member is used.  lo_server_recv_raw_stream() walks through the
array backward, closing any detected closed connections, and returning
the first message received on a socket.  Thus, new connections have
the lowest priority.

git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@148 8a0476ba-7714-0410-bd95-fb3ea4a72f84