Conditionally avoid the use of variable-argument macros if not GCC.
The use of variable-argument macros is used by LibLo to provide file
and line number information when reporting error conditions, and to
provide "markers" to check for errors in the variable argument list.
However, ANSI C compilers do not support variable-argument macros, so
this is a GCC-ism.  For non-GCC compilers, use usual variable-argument
function calls instead.

git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@157 8a0476ba-7714-0410-bd95-fb3ea4a72f84