Improve Windows support.
This patch improves Windows support in the following ways:

  - Implemention of network adapter discovery and selection for
    multicast servers and addresses using native Windows calls.

  - Compiles to a DLL using MingW as a cross-compiler on Linux.

  - MingW build now uses the .def file to ensure that export ordinals
    remain consistent with the Visual Studio build.

  - Function stubs used to call the "internal" functions (lo_send,
    etc.) can replace vararg macros with a non-GCC compiler is used.
    This was previously done for Visual Studio builds, but it is now
    done using an explicit macro and set during MingW builds, so that
    the generated MingW DLL is compatible for linking with Visual
    Studio.

  - Additional fixes to make headers work in Visual Studio, including
    removal of the "inline" keyword.

In short this patch allows a Linux host to build a Windows DLL that is
usable from Visual Studio.

To build the DLL this way, you must first build the pthreads-win32
port (http://sourceware.org/pthreads-win32/).  I created a folder
$HOME/.win with subfolders "include", "lib", "bin", and placed the
pthread headers, .a, and .dll in them, respectively.  I was then able
to successfully built the liblo DLL with the following configuration:

$ ./autogen.sh --prefix=$HOME/.win --host i586-mingw32msvc \
     CFLAGS=-I$HOME/.win/include LDFLAGS=-L$HOME/.win/lib
$ make

Note that it does _not_ build in Visual Studio 2010, although it does
build in VS 2008.  (Yay regressions!)  However, the headers and
MingW-produced DLL are usable from a VS 2010 project.

git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@190 8a0476ba-7714-0410-bd95-fb3ea4a72f84