Detect Windows Server 2003 or later, and use IP_ADD_MEMBERSHIP accordingly.
This reconciles some changes in the previous Windows-related patch
which removed the #ifdef used to differentiate multicast behaviour
during server setup on Windows and other operating systems.

The story is that I had found that on Windows it was necessary to join
the multicast group after the call to bind(), whereas on Linux and OS
X it seemed to work only if before bind().  Later, testing on Windows
Vista showed that this was no longer the case, so I removed it.
However, more testing has revealed that to retain Windows XP support
it is nonetheless necessary to keep the old before-bind() behaviour.
This an annoying but simple difference and I see no reason to break
previously working programs on Windows XP, so this patch adds a
run-time test for the Windows version that joins the multicast group
before or after bind(), accordingly.  Currently it checks for Windows
Server 2003 or later, meaning Windows version 5.2.  Windows XP is
version 5.1.

git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@192 8a0476ba-7714-0410-bd95-fb3ea4a72f84