Accept new connections in lo_server_wait(), block until data ready.
lo_server_wait() may return if a new connection is requested on the
listening socket, even if no actual data is ready to be read.  This
had the consequence of also making lo_server_recv_noblock() think that
data was ready, and block indefinitely.  This patch makes
lo_server_wait() handle incoming connection requests, and does not
return until a data stream is actually ready to be read.

git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@201 8a0476ba-7714-0410-bd95-fb3ea4a72f84