Add support for bidirectional TCP.
This involves keeping track of connection sources for every open TCP
connection.  Therefore a "sources" array has been added to lo_server
to use for storing socket and address information of the other side,
making lo_message_get_source() return a usable lo_address in method
handlers.

The required changes are summarized:

On new TCP connection,

- Keep track of sender address provided by accept().  Must be
  associated with a server's socket list item corresponding with the
  connection.  Construct a new lo_address holding host and port of
  sockaddr returned from accept().

When receiving a TCP message,

- During dispatch, need to provide message source, so look up
  previously-saved lo_address based on receiving socket.

When sending a TCP message with lo_send_from(), will need to start
listening to that socket in the associated server, so

- If address has no socket and one is created, add it to server's
  list.  This means copying the lo_address structure so as not to
  disturb user's memory management.

This all means performing a look-up on every send and every receive.
Since socket numbers are small, we keep a socket-indexed array of
lo_address structures owned by the server.

git-svn-id: https://liblo.svn.sourceforge.net/svnroot/liblo/trunk@217 8a0476ba-7714-0410-bd95-fb3ea4a72f84