Optionally perform SLIP encoding/decoding.
Discussions on OSC-dev recommend using SLIP over TCP.  We do not make
it the default, but liblo can now interoperate with other OSC
implementations that may use SLIP, as defined in RFC 1055.  This could
also be useful for communicating with a TCP-serial converter for
example.  Although a size-prefix is not used in SLIP mode, messages
are still currently limited to 32k.

A basic pattern match is used to check whether the first bytes of
incoming messages are a message path or bundle prefix, and if so, SLIP
is assumed, otherwise the first bytes are assumed to be a 32-bit
message size prefix.  Therefore incoming connections using SLIP should
work automatically, modulo the unlikely eventuality that the size
prefix bytes coincide with a possible OSC message.  Unfortunately this
detection method is not robust for streams starting at some random
position within a message, so some works needs to be done to make it
recover gracefully when partial messages are encountered.

For outgoing streams, this patch introduces a "protocol flags" field
to lo_address which may be used in the future to specify additional
transport-related options.