Remove sensitivity to trailing slash when dealing with Unix socket URLs.
Bug #2173243 reported that URLs returned by lo_address_get_url() could
not be used for with lo_address_new_from_url() in the case of Unix
sockets, because they have a trailing slash which gives the wrong path
for the named port.

Here, we remove the trailing slash in returned URLs in the case of
Unix sockets, but also remove it after parsing the port path if it is
present.  (Which should be okay since Unix socket paths cannot be
directories, therefore cannot end with a slash.)