Fix non-blocking semantics of lo_server_recv_noblock() for TCP.
This patch adds a small heap buffer to each open socket that expands
as more room is needed to read larger messages.  The buffer is used to
resume a partially-read message instead of requiring a blocking recv()
loop.

A test is included for checking that blocking doesn't occur if a
partial message is received.