Enforce ownership semantics statically by a bool template for Message and Bundle.
For instance, the compiler can now verify that Bundle only takes
ownership over messages that do not have ownership.  Otherwise
double-frees can easily occur when constructing temporary Messages on
the stack and passing them into Bundles.

It would be better to have a concept of ownership *transfer*, like
std::unique_ptr.

The syntax is not as clean due to this template argument, but forcing
the user to think about ownership is probably overall a good thing,
since it can easily lead to leaks and crashes.