Add a README to the build directory, exclude premake4.exe binary from tarball.
It doesn't fly with the Debian maintainers to include a binary.
However, this helps with building for Visual Studio, so unfortunately
VS users will now have to get premake4.exe from premake's website.  A
work-around would be to pre-generate the VS project files, but the
premake script actually does more than this--it generates config.h and
other files, so we can't do without it.  Therefore, at least a README
has been added to explain this as well as the Android build procedure.