Update C++ bindings to use liblo's reference counting in messages and bundles.
This patch also removes the template arguments previously used to
track ownership, since they are not needed in the presence of ref
counting.  A run-time variable was added to lo::Address to track
ownership in that case, since it is not ref counted.

The default behaviour of (void) handlers has been changed to `0' (stop
processing) in order to adhere to the principle of least surprise.
Users wanting to have multiple handlers trigger on a message should
explicitly "return 1" from their handlers.