If version.c does not include lo.h, it is undefined when compiled as C++.
Probably a weird symbol thing, it can be fixed by appending `extern
"C"`, but this is a weird, obscure problem anyways, only occurring if
liblo is compiled as if it were C++ code, so I'd prefer not to add
that. But in any case, it shouldn't hurt to include the library
headers here.