Add two callbacks to the server thread, one for initialization and one for cleanup.
Background: I use liblo together with portaudio to run an audio
server, to which I can connect from my main application. To function
correctly, portaudio needs to be initialized in the same thread as the
audio stream is later started in. Since the audio stream is started
from an OSC message handler, i.e. in the liblo server thread, the
portaudio initialization must happen in that same thread. But that is
not easily done, as there is no simple way to hook into the server
thread. I could of course connect to the server from my main thread
and call an init function, but that would not be very elegant.

Instead, I just added two callbacks to the server thread, one for
initialization and one for cleanup. Attaching a patch below!