Support receiving double-ENDed SLIP packets
In the rfc1055 specification it is mentioned (page two, second paragraph
under "Protocol") that senders could place an END character at the
beginning as well as the end of a SLIP encoded packet.

As it stands: when parsing such a double-ENDed packet, liblo determines
(through multiple calls to `slip_decode()`) that there are two messages
 - the first having a length of zero - and stores both into its buffer.

Upon reading the messages back out (`lo_server_buffer_copy_for_dispatch()`)
the first message having a length of zero causes the code to assume
there are no more messages, causing the second message to be skipped and
thus never dispatched to an awaiting application.

This patch resolves that issue by skipping the creation of a message
header (and the increasing of the buffer offsets) if the length of a
SLIP message is 0.