Update pointer `buffer_after` after realloc of `sc->buffer`
When created, `buffer_after` is a pointer to the same `char`-array as the one
that `sc->buffer` points to, albeit to a different point within it.

If the length of the `char`-array is deemed too small, it is increased via a
`realloc()`, which typically relocates the array elsewhere (usually later)
within memory.

Unfortunately, prior to this commit, `buffer_after` was not updated to take into
account this move.

As a result, on the next iteration of the encapsulating `while` loop, when the
pointer arithmetic `buffer_after - sc->buffer` takes place the result is either:
(a) a number that exceeds the length of the array, or (b) negative.

As the outcome of this is used to determine the position at which to start
writing the next incoming message into the `char`-array, either answer causes
a segfault.