Redefine opaque types as pointers to structs.
Warning: This may break source compatibility!

This fixes casting-related errors in the C++ wrapper related to
passing a lo::ServerThread object to lo::Address::send_from, see
https://github.com/radarsat1/liblo/issues/101

To summarize, the C++ template cannot differentiate between lo_server
and lo_server_thread because they are both (void*), and therefore
performs the wrong cast.  This can be fixed by making these opaque
types pointers to different types instead of all pointing to void.

It could be possible to redefine just these types, but this patch
takes this to the full extent and redefines all the opaque pointers so
that they are different types, to avoid any potential similar issues.

It is more correct, but could lead to incompatibility with older code,
particularly because it redefines the lo_method_handler prototype.