WIN32: detect target correctly.
When compiling the tests of RTOSC submodule of  ZynAddSubbFx, I got the same error signaled into issue #114.
This happens because  WIN32 is undefined. The right symbol to use is _WIN32 since it is intrinsic to the compiler. See:

https://docs.microsoft.com/en-us/cpp/preprocessor/predefined-macros?redirectedfrom=MSDN&view=msvc-170

Actually, testing for _MSC_VER has also no much sense here , because if _WIN32 does not exist then you surely cannot use or include WinSock.