WIN32: the nfds parameter of select() is unused.
Winsock provides an opaque SOCKET type for handling sockets and the nfds parameter of select() is ignored but kept for compatibity with BSD standard.
Expecting to handle sockets by not assuming to be an int type, this could cause a complain from MSVC when compiling for 64bit for Windows because incompatible types, so it would be better to just ignore this and just leave nfds to zero.